+++ sources/github-source/test/common.test.ts
@@ -1,9 +1,9 @@
-import {StreamBase} from '../src/streams/common';
+import {calculateUpdatedStreamState} from 'faros-airbyte-cdk';
 
 describe('calculateUpdatedStreamState', () => {
   test('should return currentStreamState if latestRecordCutoff is null', () => {
     const currentStreamState = {'github/repo1': {cutoff: 1627641720000}};
-    const result = StreamBase.calculateUpdatedStreamState(
+    const result = calculateUpdatedStreamState(
       null,
       currentStreamState,
       'github/repo1'
@@ -16,7 +16,7 @@ describe('calculateUpdatedStreamState', () => {
       'github/repo1': {cutoff: new Date('2021-07-30T00:00:00Z').getTime()},
     };
     const latestRecordCutoff = new Date('2021-07-29T00:00:00Z');
-    const result = StreamBase.calculateUpdatedStreamState(
+    const result = calculateUpdatedStreamState(
       latestRecordCutoff,
       currentStreamState,
       'github/repo1'
@@ -29,7 +29,7 @@ describe('calculateUpdatedStreamState', () => {
       'github/repo1': {cutoff: new Date('2021-07-30T00:00:00Z').getTime()},
     };
     const latestRecordCutoff = new Date('2021-08-30T00:00:00Z');
-    const result = StreamBase.calculateUpdatedStreamState(
+    const result = calculateUpdatedStreamState(
       latestRecordCutoff,
       currentStreamState,
       'github/repo1'
